requirements.dev.txt
-r requirements.common.txt
django-debug-toolbar==1.8